requirements.txt
-i https://pypi.org/simple
certifi==2025.10.5; python_version >= '3.7'
charset-normalizer==3.4.4; python_version >= '3.7'
idna==3.11; python_version >= '3.8'
pillow==11.3.0; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
